// packages/job-scheduling/src/utils/types.ts
export type TCatalystCron =
	| ICatalystOneTimeCron
	| ICatalystEveryCron
	| ICatalystDailyCron
	| ICatalystMonthlyCron
	| ICatalystYearlyCron
	| ICatalystCronExpression;

/** Details of the cron */
export interface ICatalystCronDetails extends ICatalystGResponse {
	/** Id of the cron */
	id: string;
	/** Name of the cron */
	cron_name: string;
	/** Description of the cron */
	description?: string;
	/** Type of the cron */
	cron_type: CRON_TYPE;
	/** Id of the function associated with the cron if the target type is Function */
	cron_function_id?: string;
	/** Execution type of the cron */
	cron_execution_type: CRON_EXECUTION_TYPE;
	/** Status denotes if the cron is active or not */
	cron_status: boolean;
	/** The time when the cron should stop submitting jobs to the job pool */
	end_time: string;
	/** UNIX cron expression */
	cron_expression?: string;
	/** Details of the cron */
	cron_detail: TCatalystCron;
	/** Meta details of the job */
	job_meta: TCatalystJobMetaDetails;
}

export interface ICatalystCronBasic<T extends TCatalystJobs> extends ICatalystJSON {
	/** Name of the cron */
	cron_name: string;
	/** Description of the cron */
	description?: string;
	/** Status denotes if the cron is active or not */
	cron_status: boolean;
	/** Meta details fo the job */
	job_meta: T;
}

/**
 * Cron that's execute only once
 */
export interface ICatalystOneTimeCron<T extends TCatalystJobs = TCatalystJobs>
	extends ICatalystCronBasic<T> {
	/** Cron type for OneTime cron */
	cron_type: 'OneTime';
	/** Details of the OneTime cron */
	cron_detail: {
		/** time when the cron should be executed. value should be in milliseconds */
		time_of_execution: string;
		/** timezone which should be used */
		timezone?: string;
	};
}

/**
 * A Cron that's execute for every time interval identified by the below params
 */
export interface ICatalystEveryCron<T extends TCatalystJobs = TCatalystJobs>
	extends ICatalystCronBasic<T> {
	/** Cron type for Every cron */
	cron_type: 'Periodic';
	/** Details of the every cron */
	cron_detail: {
		/** Periodicity in hours */
		hour: number;
		/** Periodicity in minutes  */
		minute: number;
		/** Periodicity in seconds */
		second: number;
		/** Timezone the cron should use */
		timezone?: string;
		/** Type of periodic cron repetition */
		repetition_type: 'every';
	};
	/** the time when the cron should stop submitting job to the job pool */
	end_time?: string;
}

/**
 * Cron that's executed daily at a particular time.
 */
export interface ICatalystDailyCron<T extends TCatalystJobs = TCatalystJobs>
	extends ICatalystCronBasic<T> {
	/** Cron type of Daily cron */
	cron_type: 'Calendar';
	/** Details of every cron */
	cron_detail: {
		/** Hour of the day. Supported values from 0-23 */
		hour: number;
		/** Minute of the hour. Supported values from 0-59 */
		minute: number;
		/** seconds of the minute. Supported values from 0-59 */
		second: number;
		/** Timezone the cron should use */
		timezone?: string;
		/** Type of periodic cron repetition for daily cron */
		repetition_type: 'daily';
	};
	/** the time when the cron should stop submitting job to the job pool */
	end_time?: string;
}

/**
 * Cron that's executed monthly
 */
export interface ICatalystMonthlyCron<T extends TCatalystJobs = TCatalystJobs>
	extends ICatalystCronBasic<T> {
	/** Cron type for monthly cron */
	cron_type: 'Calendar';
	/** Details of monthly cron */
	cron_detail: {
		/** Type of periodic cron repetition for monthly cron */
		repetition_type: 'monthly';
		/** Calendar days of the month. Supported values 1-31 */
		days?: Array<number>;
		/** Calendar week of the month. Supported values 1-5 */
		weeks_of_month?: Array<number>;
		/** Calendar day of the week. Supported values 1-7 */
		week_day?: Array<number>;
		/** Hour of the day. Supported values from 0-23 */
		hour: number;
		/** Minute of the hour. Supported values from 0-59 */
		minute: number;
		/** seconds of the minute. Supported values from 0-59 */
		second: number;
		/** Timezone the cron should use */
		timezone?: string;
	};
	/** the time when the cron should stop submitting job to the job pool */
	end_time?: string;
}

/**
 * Cron that's executed yearly
 */
export interface ICatalystYearlyCron<T extends TCatalystJobs = TCatalystJobs>
	extends ICatalystCronBasic<T> {
	/** Cron type for yearly cron */
	cron_type: 'Calendar';
	/** Details of yearly cron */
	cron_detail: {
		/** Type of periodic cron repetition for yearly cron */
		repetition_type: 'yearly';
		/** Calendar months of a year. Supported values 1-12 */
		months: Array<number>;
		/** Calendar week of the month. Supported values 1-5 */
		weeks_of_month?: Array<number>;
		/** Calendar day of the week. Supported values 1-7 */
		week_day?: Array<number>;
		/** Calendar days of the month. Supported values 1-31 */
		days?: Array<number>;
		/** Hour of the day. Supported values from 0-23 */
		hour: number;
		/** Minute of the hour. Supported values from 0-59 */
		minute: number;
		/** seconds of the minute. Supported values from 0-59 */
		second: number;
		/** Timezone the cron should use */
		timezone?: string;
	};
	/** the time when the cron should stop submitting job to the job pool */
	end_time?: string;
}

/**
 * Cron that's created based on the Linux based cron expressions.
 */
export interface ICatalystCronExpression<T extends TCatalystJobs = TCatalystJobs>
	extends ICatalystCronBasic<T> {
	/** Cron type for cron that uses Cron expression */
	cron_type: 'CronExpression';
	/** UNIX cron expression */
	cron_expression: string;
	/** Details of the expression cron */
	cron_detail: {
		/** Timezone the cron should use */
		timezone?: string;
	};
	/** the time when the cron should stop submitting job to the job pool */
	end_time?: string;
}

// JOBS
export type TCatalystCapacityAttributes = Record<CAPACITY_ATTRIBUTE_TYPE, string | number>;

/**
 * Details of the jobpool
 */
export interface ICatalystJobpoolDetails extends ICatalystGResponse {
	/** Jobpool Id */
	id: string;
	/** Type of jobpool */
	type: TARGET_TYPE;
	/** Name of the jobpool */
	name: string;
	/** Configured capacity of the jobpool */
	capacity: TCatalystCapacityAttributes;
}

/**
 * Meta details of the job
 */
export interface ICatalystJobMetaConfig {
	/** Number of time the job can be retied incase of failure */
	number_of_retries: number;
	/** Interval in which the job should be retried */
	retry_interval: number;
}

/**
 * Details of the job's target
 */
export interface ICatalystJobTargetDetails {
	/** Id of the job target */
	id: string;
	/** Name of the job target */
	target_name: string;
	/** Details of the job target */
	details: Record<string, unknown>;
}

/**
 * Details of the job's source
 */
export interface ICatalystJobSourceDetails {
	/** Id of the job's source */
	id: string;
	/** Name of the job's source */
	source_name: string;
	/** Details of the job's source */
	source_details: Record<string, unknown>;
}
// ...
export type TCatalystJobs =
	| ICatalystFunctionJob
	| ICatalystWebhookJob
	| ICatalystAppSailJob
	| ICatalystCircuitJob;

/**
 * Meta details of the job from server
 */
export type TCatalystJobMetaDetails<T = TCatalystJobs> = {
	/** Id of the job meta */
	id: string;
	/** Details of the job's target */
	target_details: ICatalystJobTargetDetails;
	/** Source type of the job */
	source_type: string;
	/** Source Id of the job */
	source_id: string;
	/** Source Details of the job */
	source_details: ICatalystJobSourceDetails;
	/** Id of the associated jobpool */
	jobpool_id: string;
	/** Details of the associated jobpool */
	jobpool_details: ICatalystJobpoolDetails;
} & T &
	ICatalystJSON;